v1.4.2 Add Jest unit testing, enhance staff commands and error handling
This update introduces Jest for unit testing, optimizes the handling of staff commands, and improves error handling. The handling of staff commands was refactored for better readability and manageability, shifting it into a separate utility file caller helperFunctions.js. The error handling was improved to facilitate better notifications to the bot owner whenever an error occurs, improving the maintainability of the application.