v1.4.5 Add manager role to team and update removal logic
This commit mainly adds the ability for a staff to assign a manager role to a team, updating both the team model and adding a new staff command. This commit also refactors the logic for removing a player from a team so that it no longer relies on the captain's ID. In addition, a new database utility to retrieve a team by its name has been implemented. A bot test file has been included as well.